+++ packages/delir-core/src/Engine/Renderer/P5js/specFixtures/globals.js
