+++ packages/delir-core/src/Engine/Renderer/RendererBase.ts
